+++ src/components/sections/ProductGallery/ProductGallery.tsx
@@ -1,21 +1,16 @@
-import { usePagination, useSearch as useFSSearch } from '@faststore/sdk'
-import { GatsbySeo } from 'gatsby-plugin-next-seo'
 import { lazy, useState } from 'react'
 import Sort from 'src/components/search/Sort'
 import ProductGridSkeleton from 'src/components/skeletons/ProductGridSkeleton'
 import SkeletonElement from 'src/components/skeletons/SkeletonElement'
-import Button, { ButtonLink } from 'src/components/ui/Button'
+import Button from 'src/components/ui/Button'
 import Icon from 'src/components/ui/Icon'
 import { mark } from 'src/sdk/tests/mark'
 import useSearch from 'src/contexts/SearchContext/useSearch'
 import Filters from 'src/components/search/PLPFilters'
-import ProductNotFound from 'src/components/common/ProductNotFound/ProductNotFound'
 
 import Section from '../Section'
-import { useProductsPrefetch } from './usePageProducts'
 import './product-gallery.scss'
-import { useGalleryQuery } from './useGalleryQuery'
-import { useDelayedFacets } from './useDelayedFacets'
+import ProductGalleryPaginator from './ProductGalleryPaginator'
 
 const GalleryPage = lazy(() => import('./ProductGalleryPage'))
 
@@ -37,29 +32,10 @@ function ProductGallery({
   galleryTitle,
   hasFilter = true,
 }: Props) {
-  const { products, productsCount, isLoading, slug } = useSearch()
+  const { products, productsCount, isLoading, slug, lastPage, currentPage } =
+    useSearch()
 
   const [isFilterOpen, setIsFilterOpen] = useState<boolean>(false)
-  const { pages, addNextPage, addPrevPage, state: searchState } = useFSSearch()
-
-  const { data } = useGalleryQuery()
-
-  const facets = useDelayedFacets(data)
-
-  const totalCount = data?.search.products.pageInfo.totalCount ?? 0
-  const { next, prev } = usePagination(totalCount)
-
-  useProductsPrefetch(prev ? prev.cursor : null)
-  useProductsPrefetch(next ? next.cursor : null)
-
-  if ((!products || !products.length) && !isLoading) {
-    return (
-      <ProductNotFound
-        title="OPS!!!!!"
-        subtitle={`Produtos de "${slug}" não encontrados.`}
-      />
-    )
-  }
 
   return (
     <Section
@@ -69,7 +45,7 @@ function ProductGallery({
       {searchTerm && (
         <header className="product-listing__search-term layout__content">
           <h1>
-            Showing results for: <span>{searchTerm}</span>
+            Mostrando resultados de: <span>{searchTerm}</span>
           </h1>
         </header>
       )}
@@ -85,6 +61,7 @@ function ProductGallery({
           <div className="product-listing__filters">
             <Filters
               slug={slug}
+              term={searchTerm}
               isFilterOpen={isFilterOpen}
               onDismiss={() => setIsFilterOpen(false)}
             />
@@ -95,18 +72,16 @@ function ProductGallery({
           className="product-listing__results-count"
           data-count={productsCount}
         >
-          <SkeletonElement shimmer type="text" loading={!data || isLoading}>
+          <SkeletonElement shimmer type="text" loading={!products || isLoading}>
             <h2 data-testid="total-product-count">
-              <span>Monstrando</span>
+              <span>Mostrando</span>
               <span>{productsCount} produtos</span>
             </h2>
           </SkeletonElement>
         </div>
 
         <div className="product-listing__sort">
-          <SkeletonElement shimmer type="text" loading={facets?.length === 0}>
-            <Sort />
-          </SkeletonElement>
+          <Sort />
 
           <SkeletonElement shimmer type="button" loading={isLoading}>
             <Button
@@ -122,65 +97,21 @@ function ProductGallery({
         </div>
 
         <div className="product-listing__results">
-          {/* Add link to previous page. This helps on SEO */}
-          {prev !== false && (
-            <div className="product-listing__pagination product-listing__pagination--top">
-              <GatsbySeo defer linkTags={[{ rel: 'prev', href: prev.link }]} />
-              <ButtonLink
-                onClick={(e) => {
-                  e.currentTarget.blur()
-                  e.preventDefault()
-                  addPrevPage()
-                }}
-                to={prev.link}
-                rel="prev"
-                variant="secondary"
-                iconPosition="left"
-                icon={
-                  <Icon name="ArrowLeft" width={16} height={16} weight="bold" />
-                }
-              >
-                Previous Page
-              </ButtonLink>
-            </div>
-          )}
           {/* Render ALL products */}
           <ProductGridSkeleton loading={!products || isLoading}>
-            {pages.map((page) => (
-              <>
-                {products?.length && (
-                  <GalleryPage
-                    key={`gallery-page-${page}`}
-                    showSponsoredProducts={false}
-                    fallbackData={
-                      page === searchState.page ? products : undefined
-                    }
-                    page={page}
-                    title={title}
-                    products={products}
-                  />
-                )}
-              </>
-            ))}
+            <GalleryPage
+              showSponsoredProducts={false}
+              page={currentPage}
+              title={title}
+              products={products}
+            />
           </ProductGridSkeleton>
           {/* Add link to next page. This helps on SEO */}
-          {next !== false && products && products.length ? (
-            <div className="product-listing__pagination product-listing__pagination--bottom">
-              <GatsbySeo defer linkTags={[{ rel: 'next', href: next.link }]} />
-              <ButtonLink
-                data-testid="show-more"
-                onClick={(e) => {
-                  e.currentTarget.blur()
-                  e.preventDefault()
-                  addNextPage()
-                }}
-                to={next.link}
-                rel="next"
-                variant="secondary"
-              >
-                CARREGAR MAIS
-              </ButtonLink>
-            </div>
+          {!isLoading && products?.length ? (
+            <ProductGalleryPaginator
+              currentPage={currentPage}
+              lastPage={lastPage}
+            />
           ) : (
             <></>
           )}